+++ src/test/scala/ClientSpec.scala
@@ -1,8 +1,7 @@
+package io.keen.client.scala
 package test
 
-import scala.collection.concurrent.TrieMap
 import scala.collection.JavaConversions._
-import scala.collection.mutable.ListBuffer
 import scala.concurrent.duration._
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.{ Await, Future, TimeoutException }
@@ -12,17 +11,11 @@ import akka.actor.ActorSystem
 import akka.pattern.AskTimeoutException
 import com.typesafe.config.{ Config, ConfigFactory }
 import org.specs2.mutable.Specification
-import org.specs2.time.NoTimeConversions
 import spray.http.Uri
 import spray.http.Uri._
 
-import io.keen.client.scala._
-
-class ClientSpec extends Specification with NoTimeConversions {
-  // Timeout used for most future awaits, etc. With unit tests/mocking this
-  // shouldn't normally need to be long.
-  val timeout = 1.second
-
+/** Fakes and configuration fixtures for Client specs */
+trait ClientSpecification extends Specification {
   class OkHttpAdapter extends HttpAdapter {
     var lastUrl: Option[String] = None
     var lastKey: Option[String] = None
@@ -37,15 +30,9 @@ class ClientSpec extends Specification with NoTimeConversions {
       params: Map[String, Option[String]] = Map.empty
     ): Future[Response] = {
 
-      // We have a map of str,opt[str] and we need to convert it to
-      val filteredParams = params.filter(
-        // Filter out keys that are None
-        _._2.isDefined
-      ).map(
-        // Convert the remaining tuples to str,str
-        param => (param._1 -> param._2.get)
-      )
-      // Make a Uri
+      // Query.apply won't accept the Option wrapper in our param values
+      val filteredParams = params.collect { case (key, Some(value)) => (key, value) }.toMap
+
       val finalUrl = Uri(
         scheme = scheme,
         authority = Authority(host = Host(authority)),
@@ -105,15 +92,12 @@ class ClientSpec extends Specification with NoTimeConversions {
       "keen.optional.write-key" -> "writeKey"
     )
   ).withFallback(defaultConfig)
+}
 
-  // generates n test events
-  def generateTestEvents(n: Integer): ListBuffer[String] = {
-    val events = ListBuffer.empty[String]
-    for (i <- 1 to n) {
-      events += s"""{"param$i":"value$i"}"""
-    }
-    events
-  }
+class ClientSpec extends ClientSpecification {
+  // Timeout used for most future awaits, etc. With unit tests/mocking this
+  // shouldn't normally need to be long.
+  val timeout = 1.second
 
   // Sequential because it's less work to share the client instance
   // TODO: set up separate read-only client, writer client, etc. instead of
@@ -282,156 +266,6 @@ class ClientSpec extends Specification with NoTimeConversions {
     }
   }
 
-  "Client with interval-based queueing enabled" should {
-    val queueConfig = ConfigFactory.parseMap(
-      Map(
-        "keen.optional.environment" -> "test",
-        "keen.queue.batch.size" -> 5,
-        "keen.queue.batch.timeout" -> "5 seconds",
-        "keen.queue.max-events-per-collection" -> 250,
-        "keen.queue.send-interval.events" -> 100,
-        "keen.queue.send-interval.duration" -> "2 seconds",
-        "keen.queue.shutdown-delay" -> "0s"
-      )
-    ).withFallback(dummyConfig)
-
-    val client = new Client(config = queueConfig) with Master {
-      override val httpAdapter = new OkHttpAdapter
-    }
-
-    val collection = "foo"
-    val projectId = client.settings.projectId
-
-    var handleMap = TrieMap.empty[String, ListBuffer[Long]]
-    val store = client.eventStore
-    var testEvents = ListBuffer.empty[String]
-
-    "send queued events" in {
-      testEvents = generateTestEvents(5)
-      testEvents.foreach { event => client.queueEvent(collection, event) }
-
-      // verify that the expected number of events are in the store
-      handleMap = store.getHandles(projectId)
-      handleMap.size must beEqualTo(1)
-      handleMap.getOrElse(collection, null).size must beEqualTo(5)
-
-      // send the queued events
-      client.sendQueuedEvents()
-
-      // validate that the store is now empty
-      handleMap = store.getHandles(projectId)
-      handleMap.size must beEqualTo(0)
-
-      // try sending events again, nothing should happen because the queue is empty
-      client.sendQueuedEvents()
-
-      // the store should still be empty
-      handleMap = store.getHandles(projectId)
-      handleMap.size must beEqualTo(0)
-    }
-
-    "automatically send queued events when queue reaches keen.queue.send-interval.events" in {
-      testEvents = generateTestEvents(100)
-
-      // queue the first 50 events
-      for (i <- 0 to 49) {
-        client.queueEvent(collection, testEvents(i))
-      }
-
-      // verify that the expected number of events are in the store
-      handleMap = store.getHandles(projectId)
-      handleMap.size must beEqualTo(1)
-      handleMap.getOrElse(collection, null).size must beEqualTo(50)
-
-      // add the final 50 events
-      for (i <- 50 to 99) {
-        client.queueEvent(collection, testEvents(i))
-      }
-
-      // validate that the store is now empty as a result of sendQueuedEvents being automatically
-      // triggered with the queueing of the 100th event
-      handleMap = store.getHandles(projectId)
-      handleMap.size must beEqualTo(0)
-    }
-
-    "automatically send queued events every keen.queue.send-interval.duration" in {
-      testEvents = generateTestEvents(5)
-      testEvents.foreach { event => client.queueEvent(collection, event) }
-
-      // verify that the expected number of events are in the store
-      handleMap = store.getHandles(projectId)
-      handleMap.size must beEqualTo(1)
-      handleMap.getOrElse(collection, null).size must beEqualTo(5)
-
-      // sleep until the set interval is reached
-      Thread.sleep((client.settings.sendIntervalDuration + 100.millis).toMillis)
-
-      // validate that the store is now empty as a result of sendQueuedEvents being automatically
-      // triggered with the queueing of the 100th event
-      handleMap = store.getHandles(projectId)
-      handleMap.size must beEqualTo(0)
-    }
-
-    "send queued events on shutdown" in {
-      testEvents = generateTestEvents(5)
-      testEvents.foreach { event => client.queueEvent(collection, event) }
-
-      // verify that the expected number of events are in the store
-      handleMap = store.getHandles(projectId)
-      handleMap.size must beEqualTo(1)
-      handleMap.getOrElse(collection, null).size must beEqualTo(5)
-
-      // send the queued events
-      client.shutdown()
-
-      // validate that the store is now empty
-      handleMap = store.getHandles(projectId)
-      handleMap.size must beEqualTo(0)
-    }
-  }
-
-  "Client with simple queueing enabled" should {
-    val queueConfig = ConfigFactory.parseMap(
-      Map(
-        "keen.optional.environment" -> "test",
-        "keen.queue.batch.size" -> 5,
-        "keen.queue.batch.timeout" -> "5 seconds",
-        "keen.queue.max-events-per-collection" -> 250,
-        "keen.queue.send-interval.events" -> 0,
-        "keen.queue.send-interval.duration" -> "0s",
-        "keen.queue.shutdown-delay" -> "0s"
-      )
-    ).withFallback(dummyConfig)
-
-    val client = new Client(config = queueConfig) with Master {
-      override val httpAdapter = new OkHttpAdapter
-    }
-
-    val collection = "foo"
-    val projectId = client.settings.projectId
-
-    var handleMap = TrieMap.empty[String, ListBuffer[Long]]
-    val store = client.eventStore
-    var testEvents = ListBuffer.empty[String]
-
-    "not exceed keen.queue.max-events-per-collection" in {
-      testEvents = generateTestEvents(500)
-      testEvents.foreach { event => client.queueEvent(collection, event) }
-
-      // verify that the expected number of events are in the store
-      handleMap = store.getHandles(projectId)
-      handleMap.size must beEqualTo(1)
-      handleMap.getOrElse(collection, null).size must beEqualTo(store.maxEventsPerCollection)
-
-      // shutdown the client
-      client.shutdown()
-
-      // validate that the store is now empty
-      handleMap = store.getHandles(projectId)
-      handleMap.size must beEqualTo(0)
-    }
-  }
-
   "Client with Spray HttpAdapter" should {
     lazy val externalSystem = ActorSystem("keen-test-user-supplied")
 
@@ -474,35 +308,6 @@ class ClientSpec extends Specification with NoTimeConversions {
     "handle 500" in {
       val res = Await.result(client.getProjects, timeout)
       res.statusCode must beEqualTo(500)
-    }
-
-    "send queued events with server failure" in {
-      val projectId = client.settings.projectId
-      val collection = "foo"
-      val testEvents = generateTestEvents(5)
-
-      // queue the events
-      for (event <- testEvents) {
-        client.queueEvent(collection, event)
-      }
-
-      // verify that the expected number of events are in the store
-      val store = client.eventStore
-      var handleMap = store.getHandles(projectId)
-      handleMap.size must beEqualTo(1)
-      handleMap.getOrElse(collection, null).size must beEqualTo(5)
-
-      // send the queued events
-      client.sendQueuedEvents()
-
-      // validate that the store still contains all of the queued events
-      handleMap = store.getHandles(projectId)
-      handleMap.size must beEqualTo(1)
-      handleMap.getOrElse(collection, null).size must beEqualTo(5)
-
-      // shutdown the client
-      client.shutdown()
-      true must beEqualTo(true)
     }
   }